inkluder begrunnelseRedusert og bruttoUtbetalt i mapping fra Simba-format
