+++ src/test/kotlin/no/nav/syfo/simba/MapInntektsmeldingFraSimbaTest.kt
@@ -1,0 +1,99 @@
+package no.nav.syfo.simba
+
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.BegrunnelseIngenEllerRedusertUtbetalingKode
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.FullLonnIArbeidsgiverPerioden
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.InntektsmeldingDokument
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Naturalytelse
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.NaturalytelseKode
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Periode
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.Refusjon
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.RefusjonEndring
+import no.nav.helsearbeidsgiver.felles.inntektsmelding.felles.models.ÅrsakInnsending
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+import java.math.BigDecimal
+import java.time.LocalDate
+import java.time.OffsetDateTime
+
+class MapInntektsmeldingFraSimbaTest {
+
+    @Test
+    fun mapInntektsmeldingMedNaturalytelser() {
+        val naturalytelser = NaturalytelseKode.values().map { Naturalytelse(it, LocalDate.now(), BigDecimal.ONE) }
+        val antallNaturalytelser = naturalytelser.count()
+        val imd = lagInntektsMeldingDokument().copy(naturalytelser = naturalytelser)
+        val mapped = mapInntektsmelding("1323", "sdfds", "134", imd)
+        assertEquals(antallNaturalytelser, mapped.opphørAvNaturalYtelse.size)
+        val naturalytelse = mapped.opphørAvNaturalYtelse.get(0)
+        assertEquals(no.nav.syfo.domain.inntektsmelding.Naturalytelse.AKSJERGRUNNFONDSBEVISTILUNDERKURS, naturalytelse.naturalytelse)
+    }
+
+    @Test
+    fun mapRefusjon() {
+        val refusjonEndring = listOf(RefusjonEndring(123.toBigDecimal(), LocalDate.now()))
+        val refusjon = Refusjon(true, BigDecimal.TEN, LocalDate.of(2025, 12, 12), refusjonEndring)
+        val mapped = mapInntektsmelding("1323", "sdfds", "134", lagInntektsMeldingDokument().copy(refusjon = refusjon))
+        assertEquals(mapped.refusjon.opphoersdato, refusjon.refusjonOpphører)
+        assertEquals(mapped.endringerIRefusjon.size, 1)
+    }
+
+    @Test
+    fun mapBegrunnelseRedusert() {
+        val begrunnelser = BegrunnelseIngenEllerRedusertUtbetalingKode.values().toList()
+        val mapped = begrunnelser.map { kode ->
+            lagInntektsMeldingDokument().copy(
+                fullLønnIArbeidsgiverPerioden = FullLonnIArbeidsgiverPerioden(
+                    false,
+                    kode,
+                    BigDecimal.ONE
+                )
+            )
+        }
+            .map { imd ->
+                mapInntektsmelding("123", "abc", "345", imd)
+            }.toList()
+        for (i in mapped.indices) {
+            assertEquals(begrunnelser[i].value, mapped[i].begrunnelseRedusert, "Feil ved mapping av $i: ${begrunnelser[i]}")
+            assertEquals(BigDecimal(1), mapped[i].bruttoUtbetalt, "Feil ved mapping av $i: ${begrunnelser[i]}")
+        }
+    }
+
+    @Test
+    fun mapIngenBegrunnelseRedusert() {
+        val im = mapInntektsmelding("1", "2", "3", lagInntektsMeldingDokument())
+        assertEquals("", im.begrunnelseRedusert)
+        assertEquals(BigDecimal(0), im.bruttoUtbetalt)
+    }
+
+    private fun lagInntektsMeldingDokument(): InntektsmeldingDokument {
+        val dato1 = LocalDate.now().minusDays(7)
+        val dato2 = LocalDate.now().minusDays(5)
+        val dato3 = LocalDate.now().minusDays(3)
+        val periode1 = listOf(Periode(dato1, dato1))
+        val periode2 = listOf(Periode(dato1, dato1), Periode(dato2, dato2))
+        val periode3 = listOf(Periode(dato1, dato3))
+
+        val refusjonEndring = listOf(RefusjonEndring(123.toBigDecimal(), LocalDate.now()))
+        val refusjon = Refusjon(true, BigDecimal.TEN, LocalDate.of(2025, 12, 12), refusjonEndring)
+
+        val imDokumentFraSimba = InntektsmeldingDokument(
+            orgnrUnderenhet = "123456789",
+            identitetsnummer = "12345678901",
+            fulltNavn = "Test testesen",
+            virksomhetNavn = "Blåbærsyltetøy A/S",
+            behandlingsdager = listOf(dato1),
+            egenmeldingsperioder = periode1,
+            bestemmendeFraværsdag = dato2,
+            fraværsperioder = periode2,
+            arbeidsgiverperioder = periode3,
+            beregnetInntekt = BigDecimal.valueOf(100_000L),
+            refusjon = refusjon,
+            naturalytelser = emptyList(),
+            tidspunkt = OffsetDateTime.now(),
+            årsakInnsending = ÅrsakInnsending.NY,
+            innsenderNavn = "Peppes Pizza",
+            telefonnummer = "22555555"
+        )
+        return imDokumentFraSimba
+    }
+}
